Split the canvas into the concerns it actually has
PreviewCanvas had grown to 760 lines carrying five separate jobs. Three of
them are now their own hooks, each independently readable and testable:

- useCanvasZoom — every way in (pointer, pinch, keyboard, buttons, gestures
  forwarded up from inside a frame) and the one way of applying it, since the
  anchoring correction has to wait for the layout the new scale produces.
- useCanvasPan — space-drag and middle-drag.
- useFrameScrollSync — the position exchange that keeps the frames in step.

What is left is geometry and rendering. Behaviour is unchanged: re-measured
pointer anchoring in a browser after the split, and it still holds its point
to the pixel on both axes, through a wheel and through a forwarded pinch.
